refactor: merge parser codegen shared
