+++ packages/@lwc/template-compiler/src/codegen/index.ts
@@ -13,24 +13,42 @@ import { ResolvedConfig } from '../config';
 
 import {
     isCommentNode,
-    isElement,
-    isCustomElement,
     isTemplate,
     isTextNode,
     isSlot,
+    isStringLiteral,
+    isForBlock,
+    isIfBlock,
+    isForEach,
+    isBaseElement,
+    isDynamicDirective,
+    isExpression,
+    isProperty,
+    getElementDirective,
+    getKeyDirective,
+    getDomDirective,
+    isComponent,
 } from '../shared/ir';
 import { TEMPLATE_PARAMS, TEMPLATE_FUNCTION_NAME } from '../shared/constants';
 import {
-    IRNode,
-    IRElement,
-    IRText,
-    IRAttribute,
-    IRAttributeType,
-    IRComment,
+    Root,
+    ParentNode,
+    ChildNode,
+    Text,
+    IfBlock,
+    ForBlock,
+    ForEach,
+    Slot,
+    Component,
+    Element,
+    Attribute,
+    Property,
+    Comment,
+    ForOf,
 } from '../shared/types';
 
 import CodeGen from './codegen';
-import { bindExpression } from './scope';
+import Scope from './scope';
 import {
     identifierFromComponentName,
     objectToAST,
@@ -41,6 +59,7 @@ import {
     parseStyleText,
     hasIdAttribute,
     styleMapToStyleDeclsAST,
+    arrayToObjectMapper,
 } from './helpers';
 
 import { format as formatModule } from './formatters/module';
@@ -53,28 +72,29 @@ import {
     isFragmentOnlyUrl,
     isIdReferencingAttribute,
     isSvgUseHref,
+    propertyToAttributeName,
 } from '../parser/attribute';
 import { SVG_NAMESPACE_URI } from '../parser/constants';
 
-function transform(codeGen: CodeGen): t.Expression {
-    const parentStack: IRNode[] = [];
-
-    function transformElement(element: IRElement): t.Expression {
+function transform(codeGen: CodeGen, scope: Scope): t.Expression {
+    function transformElement(element: Element | Component | Slot): t.Expression {
         const databag = elementDataBag(element);
         let res: t.Expression;
 
         const children = transformChildren(element);
 
         // Check wether it has the special directive lwc:dynamic
-        if (element.lwc && element.lwc.dynamic) {
-            const expression = bindExpression(element.lwc.dynamic, element, parentStack);
-            res = codeGen.genDynamicElement(element.tag, expression, databag, children);
-        } else if (isCustomElement(element)) {
+        const { name } = element;
+        const dynamic = getElementDirective(element, isDynamicDirective);
+        if (dynamic) {
+            const expression = scope.bindExpression(dynamic.value);
+            res = codeGen.genDynamicElement(name, expression, databag, children);
+        } else if (isComponent(element)) {
             // Make sure to register the component
-            const componentClassName = element.component!;
+            const componentClassName = name;
 
             res = codeGen.genCustomElement(
-                element.tag,
+                name,
                 identifierFromComponentName(componentClassName),
                 databag,
                 children
@@ -82,64 +102,39 @@ function transform(codeGen: CodeGen): t.Expression {
         } else if (isSlot(element)) {
             const defaultSlot = children;
 
-            res = codeGen.getSlot(element.slotName!, databag, defaultSlot);
+            res = codeGen.getSlot(name, databag, defaultSlot);
         } else {
-            res = codeGen.genElement(element.tag, databag, children);
+            res = codeGen.genElement(name, databag, children);
         }
-
-        res = applyInlineIf(element, res);
-        res = applyInlineFor(element, res);
 
         return res;
     }
 
-    function transformTemplate(element: IRElement): t.Expression | t.Expression[] {
-        const children = transformChildren(element);
-
-        let res = applyTemplateIf(element, children);
-
-        if (element.forEach) {
-            res = applyTemplateFor(element, res);
-        } else if (element.forOf) {
-            res = applyTemplateForOf(element, res);
-        }
-
-        if (t.isArrayExpression(res) && element.if) {
-            // The `if` transformation does not use the SpreadElement, neither null, therefore we can safely
-            // typecast it to t.Expression[]
-            return res.elements as t.Expression[];
-        } else {
-            return res;
-        }
-    }
-
-    function transformText(consecutiveText: IRText[]): t.Expression {
+    function transformText(consecutiveText: Text[]): t.Expression {
         return codeGen.genText(
             consecutiveText.map((text) => {
                 const { value } = text;
 
-                return typeof value === 'string' ? value : bindExpression(value, text, parentStack);
+                return isStringLiteral(value) ? value.value : scope.bindExpression(value);
             })
         );
     }
 
-    function transformComment(comment: IRComment): t.Expression {
+    function transformComment(comment: Comment): t.Expression {
         return codeGen.genComment(comment.value);
     }
 
-    function transformChildren(parent: IRElement): t.Expression {
+    function transformChildren(parent: ParentNode): t.Expression {
         const res: t.Expression[] = [];
         const children = parent.children;
         const childrenIterator = children[Symbol.iterator]();
-        let current: IteratorResult<IRNode>;
-
-        parentStack.push(parent);
+        let current: IteratorResult<ChildNode>;
 
         while ((current = childrenIterator.next()) && !current.done) {
             let child = current.value;
 
             if (isTextNode(child)) {
-                const continuousText: IRText[] = [];
+                const continuousText: Text[] = [];
 
                 // Consume all the contiguous text nodes.
                 do {
@@ -156,12 +151,20 @@ function transform(codeGen: CodeGen): t.Expression {
                 }
             }
 
-            if (isElement(child)) {
+            if (isForBlock(child)) {
+                res.push(transformFor(child));
+            }
+
+            if (isIfBlock(child)) {
+                const ifBlockChildren = transformIf(child);
+                Array.isArray(ifBlockChildren)
+                    ? res.push(...ifBlockChildren)
+                    : res.push(ifBlockChildren);
+            }
+
+            if (isBaseElement(child)) {
                 if (isTemplate(child)) {
-                    const templateChildren = transformTemplate(child);
-                    Array.isArray(templateChildren)
-                        ? res.push(...templateChildren)
-                        : res.push(templateChildren);
+                    res.push(transformChildren(child));
                 } else {
                     res.push(transformElement(child));
                 }
@@ -172,9 +175,9 @@ function transform(codeGen: CodeGen): t.Expression {
             }
         }
 
-        parentStack.pop();
-
-        if (shouldFlatten(children, codeGen)) {
+        if (isForBlock(parent) || isIfBlock(parent)) {
+            return res[0];
+        } else if (shouldFlatten(children, codeGen)) {
             if (children.length === 1 && !containsDynamicChildren(children)) {
                 return res[0];
             } else {
@@ -185,22 +188,42 @@ function transform(codeGen: CodeGen): t.Expression {
         }
     }
 
+    function transformIf(ifBlock: IfBlock): t.Expression | t.Expression[] {
+        const children = transformChildren(ifBlock);
+
+        let res: t.Expression;
+        if (isTemplate(ifBlock)) {
+            res = applyTemplateIf(ifBlock, children);
+        } else {
+            let expression = children;
+            if (t.isArrayExpression(expression) && expression.elements.length === 1) {
+                expression = expression.elements[0] as t.Expression;
+            }
+
+            res = applyInlineIf(ifBlock, expression);
+        }
+
+        if (t.isArrayExpression(res)) {
+            // The `if` transformation does not use the SpreadElement, neither null, therefore we can safely
+            // typecast it to t.Expression[]
+            return res.elements as t.Expression[];
+        }
+
+        return res;
+    }
+
     function applyInlineIf(
-        element: IRElement,
+        ifBlock: IfBlock,
         node: t.Expression,
         testExpression?: t.Expression,
         falseValue?: t.Expression
-    ): t.Expression {
-        if (!element.if) {
-            return node;
-        }
-
+    ) {
         if (!testExpression) {
-            testExpression = bindExpression(element.if!, element, parentStack);
+            testExpression = scope.bindExpression(ifBlock.condition);
         }
 
         let leftExpression: t.Expression;
-        const modifier = element.ifModifier!;
+        const modifier = ifBlock.modifier!;
         if (modifier === 'true') {
             leftExpression = testExpression;
         } else if (modifier === 'false') {
@@ -216,18 +239,51 @@ function transform(codeGen: CodeGen): t.Expression {
         return t.conditionalExpression(leftExpression, node, falseValue ?? t.literal(null));
     }
 
-    function applyInlineFor(element: IRElement, node: t.Expression) {
-        if (!element.forEach) {
-            return node;
+    function transformFor(forBlock: ForBlock): t.Expression {
+        const children = transformForChildren(forBlock);
+
+        let expression = children;
+        if (t.isArrayExpression(expression) && expression.elements.length === 1) {
+            expression = expression.elements[0] as t.Expression;
         }
 
-        const { expression, item, index } = element.forEach;
+        let res: t.Expression;
+        if (isForEach(forBlock)) {
+            res = applyInlineFor(forBlock, expression);
+        } else {
+            res = applyInlineForOf(forBlock, expression);
+        }
+
+        return res;
+    }
+
+    function transformForChildren(forBlock: ForBlock): t.Expression {
+        scope.beginScope();
+
+        if (isForEach(forBlock)) {
+            const { item, index } = forBlock;
+            if (index) {
+                scope.declare(index);
+            }
+            scope.declare(item);
+        } else {
+            scope.declare(forBlock.iterator);
+        }
+
+        const children = transformChildren(forBlock);
+        scope.endScope();
+
+        return children;
+    }
+
+    function applyInlineFor(forEach: ForEach, node: t.Expression): t.Expression {
+        const { expression, item, index } = forEach;
         const params = [item];
         if (index) {
             params.push(index);
         }
 
-        const iterable = bindExpression(expression, element, parentStack);
+        const iterable = scope.bindExpression(expression);
         const iterationFunction = t.functionExpression(
             null,
             params,
@@ -237,12 +293,8 @@ function transform(codeGen: CodeGen): t.Expression {
         return codeGen.genIterator(iterable, iterationFunction);
     }
 
-    function applyInlineForOf(element: IRElement, node: t.Expression) {
-        if (!element.forOf) {
-            return node;
-        }
-
-        const { expression, iterator } = element.forOf;
+    function applyInlineForOf(forOf: ForOf, node: t.Expression): t.Expression {
+        const { expression, iterator } = forOf;
         const { name: iteratorName } = iterator;
 
         const argsMapping = {
@@ -259,7 +311,7 @@ function transform(codeGen: CodeGen): t.Expression {
             )
         );
 
-        const iterable = bindExpression(expression, element, parentStack);
+        const iterable = scope.bindExpression(expression);
         const iterationFunction = t.functionExpression(
             null,
             iteratorArgs,
@@ -274,152 +326,133 @@ function transform(codeGen: CodeGen): t.Expression {
         return codeGen.genIterator(iterable, iterationFunction);
     }
 
-    function applyTemplateForOf(element: IRElement, fragmentNodes: t.Expression) {
-        let expression = fragmentNodes;
-        if (t.isArrayExpression(expression) && expression.elements.length === 1) {
-            expression = expression.elements[0] as t.Expression;
-        }
-
-        return applyInlineForOf(element, expression);
-    }
-
-    function applyTemplateFor(element: IRElement, fragmentNodes: t.Expression): t.Expression {
-        let expression = fragmentNodes;
-        if (t.isArrayExpression(expression) && expression.elements.length === 1) {
-            expression = expression.elements[0] as t.Expression;
-        }
-
-        return applyInlineFor(element, expression);
-    }
-
-    function applyTemplateIf(element: IRElement, fragmentNodes: t.Expression): t.Expression {
-        if (!element.if) {
-            return fragmentNodes;
-        }
-
+    function applyTemplateIf(ifBlock: IfBlock, fragmentNodes: t.Expression): t.Expression {
         if (t.isArrayExpression(fragmentNodes)) {
             // Bind the expression once for all the template children
-            const testExpression = bindExpression(element.if!, element, parentStack);
+            const testExpression = scope.bindExpression(ifBlock.condition);
 
             return t.arrayExpression(
                 fragmentNodes.elements.map((child) =>
                     child !== null
-                        ? applyInlineIf(element, child as t.Expression, testExpression)
+                        ? applyInlineIf(ifBlock, child as t.Expression, testExpression)
                         : null
                 )
             );
         } else {
             // If the template has a single children, make sure the ternary expression returns an array
-            return applyInlineIf(element, fragmentNodes, undefined, t.arrayExpression([]));
+            return applyInlineIf(ifBlock, fragmentNodes, undefined, t.arrayExpression([]));
         }
     }
 
-    function computeAttrValue(attr: IRAttribute, element: IRElement): t.Expression {
-        const { tag, namespace } = element;
-        const isUsedAsAttribute = isAttribute(element, attr.name);
+    function computeAttrValue(
+        attr: Attribute | Property,
+        element: Element | Component | Slot
+    ): t.Expression {
+        const { name: elmName, namespace = '' } = element;
+        const { name, value: attrValue } = attr;
+        // Evaluate properties based on their attribute name
+        const attrName = isProperty(attr) ? propertyToAttributeName(name) : name;
+        const isUsedAsAttribute = isAttribute(element, attrName);
 
-        switch (attr.type) {
-            case IRAttributeType.Expression: {
-                const expression = bindExpression(attr.value, element, parentStack);
+        if (isExpression(attrValue)) {
+            const expression = scope.bindExpression(attrValue);
 
-                // TODO [#2012]: Normalize global boolean attrs values passed to custom elements as props
-                if (isUsedAsAttribute && isBooleanAttribute(attr.name, tag)) {
-                    // We need to do some manipulation to allow the diffing algorithm add/remove the attribute
-                    // without handling special cases at runtime.
-                    return codeGen.genBooleanAttributeExpr(expression);
-                }
-                if (attr.name === 'tabindex') {
-                    return codeGen.genTabIndex([expression]);
-                }
-                if (attr.name === 'id' || isIdReferencingAttribute(attr.name)) {
-                    return codeGen.genScopedId(expression);
-                }
-                if (
-                    codeGen.scopeFragmentId &&
-                    isAllowedFragOnlyUrlsXHTML(tag, attr.name, namespace)
-                ) {
-                    return codeGen.genScopedFragId(expression);
-                }
-                if (isSvgUseHref(tag, attr.name, namespace)) {
-                    codeGen.usedLwcApis.add('sanitizeAttribute');
+            // TODO [#2012]: Normalize global boolean attrs values passed to custom elements as props
+            if (isUsedAsAttribute && isBooleanAttribute(attrName, elmName)) {
+                // We need to do some manipulation to allow the diffing algorithm add/remove the attribute
+                // without handling special cases at runtime.
+                return codeGen.genBooleanAttributeExpr(expression);
+            }
+            if (attrName === 'tabindex') {
+                return codeGen.genTabIndex([expression]);
+            }
+            if (attrName === 'id' || isIdReferencingAttribute(attrName)) {
+                return codeGen.genScopedId(expression);
+            }
+            if (
+                codeGen.scopeFragmentId &&
+                isAllowedFragOnlyUrlsXHTML(elmName, attrName, namespace)
+            ) {
+                return codeGen.genScopedFragId(expression);
+            }
+            if (isSvgUseHref(elmName, attrName, namespace)) {
+                codeGen.usedLwcApis.add('sanitizeAttribute');
 
-                    return t.callExpression(t.identifier('sanitizeAttribute'), [
-                        t.literal(tag),
-                        t.literal(namespace),
-                        t.literal(attr.name),
-                        codeGen.genScopedFragId(expression),
-                    ]);
-                }
-
-                return expression;
+                return t.callExpression(t.identifier('sanitizeAttribute'), [
+                    t.literal(elmName),
+                    t.literal(namespace),
+                    t.literal(attrName),
+                    codeGen.genScopedFragId(expression),
+                ]);
             }
 
-            case IRAttributeType.String: {
-                if (attr.name === 'id') {
-                    return codeGen.genScopedId(attr.value);
-                }
-                if (attr.name === 'spellcheck') {
-                    return t.literal(attr.value.toLowerCase() !== 'false');
-                }
-
-                if (!isUsedAsAttribute && isBooleanAttribute(attr.name, tag)) {
-                    // We are in presence of a string value, for a recognized boolean attribute, which is used as
-                    // property. for these cases, always set the property to true.
-                    return t.literal(true);
-                }
-
-                if (isIdReferencingAttribute(attr.name)) {
-                    return codeGen.genScopedId(attr.value);
-                }
-                if (
-                    codeGen.scopeFragmentId &&
-                    isAllowedFragOnlyUrlsXHTML(tag, attr.name, namespace) &&
-                    isFragmentOnlyUrl(attr.value)
-                ) {
-                    return codeGen.genScopedFragId(attr.value);
-                }
-                if (isSvgUseHref(tag, attr.name, namespace)) {
-                    codeGen.usedLwcApis.add('sanitizeAttribute');
-
-                    return t.callExpression(t.identifier('sanitizeAttribute'), [
-                        t.literal(tag),
-                        t.literal(namespace),
-                        t.literal(attr.name),
-                        isFragmentOnlyUrl(attr.value)
-                            ? codeGen.genScopedFragId(attr.value)
-                            : t.literal(attr.value),
-                    ]);
-                }
-                return t.literal(attr.value);
+            return expression;
+        } else if (isStringLiteral(attrValue)) {
+            if (attrName === 'id') {
+                return codeGen.genScopedId(attrValue.value);
+            }
+            if (attrName === 'spellcheck') {
+                return t.literal(attrValue.value.toLowerCase() !== 'false');
             }
 
-            case IRAttributeType.Boolean: {
-                // A boolean value used in an attribute should always generate .setAttribute(attr.name, ''),
-                // regardless if is a boolean attribute or not.
-                return isUsedAsAttribute ? t.literal('') : t.literal(attr.value);
+            if (!isUsedAsAttribute && isBooleanAttribute(attrName, elmName)) {
+                // We are in presence of a string value, for a recognized boolean attribute, which is used as
+                // property. for these cases, always set the property to true.
+                return t.literal(true);
             }
+
+            if (isIdReferencingAttribute(attrName)) {
+                return codeGen.genScopedId(attrValue.value);
+            }
+            if (
+                codeGen.scopeFragmentId &&
+                isAllowedFragOnlyUrlsXHTML(elmName, attrName, namespace) &&
+                isFragmentOnlyUrl(attrValue.value)
+            ) {
+                return codeGen.genScopedFragId(attrValue.value);
+            }
+            if (isSvgUseHref(elmName, attrName, namespace)) {
+                codeGen.usedLwcApis.add('sanitizeAttribute');
+
+                return t.callExpression(t.identifier('sanitizeAttribute'), [
+                    t.literal(elmName),
+                    t.literal(namespace),
+                    t.literal(attrName),
+                    isFragmentOnlyUrl(attrValue.value)
+                        ? codeGen.genScopedFragId(attrValue.value)
+                        : t.literal(attrValue.value),
+                ]);
+            }
+            return t.literal(attrValue.value);
+        } else {
+            // A boolean value used in an attribute should always generate .setAttribute(attr.name, ''),
+            // regardless if is a boolean attribute or not.
+            return isUsedAsAttribute ? t.literal('') : t.literal(attrValue.value);
         }
     }
 
-    function elementDataBag(element: IRElement): t.ObjectExpression {
+    function elementDataBag(element: Element | Component | Slot): t.ObjectExpression {
         const data: t.Property[] = [];
 
-        const { attrs, props, on, forKey, lwc } = element;
+        const { attributes, properties, listeners } = element;
+        const dom = getDomDirective(element);
+        const forKey = getKeyDirective(element);
 
         // Attributes
-        if (attrs) {
+        if (attributes.length) {
             const rest: { [name: string]: t.Expression } = {};
 
-            for (const [name, value] of Object.entries(attrs)) {
+            for (const attr of attributes) {
+                const { name, value } = attr;
                 if (name === 'class') {
                     // Handle class attribute:
                     // - expression values are turned into a `className` property.
                     // - string values are parsed and turned into a `classMap` object associating
                     //   each individual class name with a `true` boolean.
-                    if (value.type === IRAttributeType.Expression) {
-                        const classExpression = bindExpression(value.value, element, parentStack);
+                    if (isExpression(value)) {
+                        const classExpression = scope.bindExpression(value);
                         data.push(t.property(t.identifier('className'), classExpression));
-                    } else if (value.type === IRAttributeType.String) {
+                    } else if (isStringLiteral(value)) {
                         const classNames = parseClassNames(value.value);
                         const classMap = t.objectExpression(
                             classNames.map((name) => t.property(t.literal(name), t.literal(true)))
@@ -431,16 +464,16 @@ function transform(codeGen: CodeGen): t.Expression {
                     // - expression values are turned into a `style` property.
                     // - string values are parsed and turned into a `styles` array
                     // containing triples of [name, value, important (optional)]
-                    if (value.type === IRAttributeType.Expression) {
-                        const styleExpression = bindExpression(value.value, element, parentStack);
+                    if (isExpression(value)) {
+                        const styleExpression = scope.bindExpression(value);
                         data.push(t.property(t.identifier('style'), styleExpression));
-                    } else if (value.type === IRAttributeType.String) {
+                    } else if (isStringLiteral(value)) {
                         const styleMap = parseStyleText(value.value);
                         const styleAST = styleMapToStyleDeclsAST(styleMap);
                         data.push(t.property(t.identifier('styleDecls'), styleAST));
                     }
                 } else {
-                    rest[name] = computeAttrValue(attrs[name], element);
+                    rest[name] = computeAttrValue(attr, element);
                 }
             }
 
@@ -453,17 +486,20 @@ function transform(codeGen: CodeGen): t.Expression {
         }
 
         // Properties
-        if (props) {
+        if (properties.length) {
+            const props = arrayToObjectMapper(properties, (prop) => prop.name);
             const propsObj = objectToAST(props, (key) => computeAttrValue(props[key], element));
             data.push(t.property(t.identifier('props'), propsObj));
         }
 
         // Context
-        if (lwc?.dom) {
+        if (dom) {
             const contextObj = t.objectExpression([
                 t.property(
                     t.identifier('lwc'),
-                    t.objectExpression([t.property(t.identifier('dom'), t.literal(lwc.dom))])
+                    t.objectExpression([
+                        t.property(t.identifier('dom'), t.literal(dom.value.value)),
+                    ])
                 ),
             ]);
             data.push(t.property(t.identifier('context'), contextObj));
@@ -472,7 +508,7 @@ function transform(codeGen: CodeGen): t.Expression {
         // Key property on VNode
         if (forKey) {
             // If element has user-supplied `key` or is in iterator, call `api.k`
-            const forKeyExpression = bindExpression(forKey, element, parentStack);
+            const forKeyExpression = scope.bindExpression(forKey.value);
             const generatedKey = codeGen.genKey(t.literal(codeGen.generateKey()), forKeyExpression);
             data.push(t.property(t.identifier('key'), generatedKey));
         } else {
@@ -482,14 +518,15 @@ function transform(codeGen: CodeGen): t.Expression {
         }
 
         // Event handler
-        if (on) {
-            const onObj = objectToAST(on, (key) => {
-                const componentHandler = bindExpression(on[key], element, parentStack);
+        if (listeners.length) {
+            const listenerObj = arrayToObjectMapper(listeners, (listener) => listener.name);
+            const onlistenerObjAST = objectToAST(listenerObj, (key) => {
+                const componentHandler = scope.bindExpression(listenerObj[key].handler);
                 const handler = codeGen.genBind(componentHandler);
 
-                return memorizeHandler(codeGen, element, parentStack, componentHandler, handler);
+                return memorizeHandler(codeGen, scope, componentHandler, handler);
             });
-            data.push(t.property(t.identifier('on'), onObj));
+            data.push(t.property(t.identifier('on'), onlistenerObjAST));
         }
 
         // SVG handling
@@ -503,8 +540,8 @@ function transform(codeGen: CodeGen): t.Expression {
     return transformChildren(codeGen.root);
 }
 
-function generateTemplateFunction(codeGen: CodeGen): t.FunctionDeclaration {
-    const returnedValue = transform(codeGen);
+function generateTemplateFunction(codeGen: CodeGen, scope: Scope): t.FunctionDeclaration {
+    const returnedValue = transform(codeGen, scope);
 
     const args = [
         TEMPLATE_PARAMS.API,
@@ -567,15 +604,16 @@ function generateTemplateFunction(codeGen: CodeGen): t.FunctionDeclaration {
     );
 }
 
-export default function (root: IRElement, config: ResolvedConfig): string {
+export default function (root: Root, config: ResolvedConfig): string {
     const scopeFragmentId = hasIdAttribute(root);
     const codeGen = new CodeGen({
         root,
         config,
         scopeFragmentId,
     });
+    const scope = new Scope();
 
-    const templateFunction = generateTemplateFunction(codeGen);
+    const templateFunction = generateTemplateFunction(codeGen, scope);
 
     let program: t.Program;
     switch (config.format) {
