+++ packages/@lwc/template-compiler/src/codegen/scope.ts
@@ -8,40 +8,66 @@ import { walk } from 'estree-walker';
 
 import * as t from '../shared/estree';
 import { TEMPLATE_PARAMS } from '../shared/constants';
-import { isComponentProp } from '../shared/ir';
-import { IRNode, TemplateExpression } from '../shared/types';
+import { Expression, Literal } from '../shared/types';
 
-/**
- * Bind the passed expression to the component instance. It applies the following transformation to the expression:
- * - {value} --> {$cmp.value}
- * - {value[index]} --> {$cmp.value[$cmp.index]}
- */
-export function bindExpression(
-    expression: TemplateExpression,
-    irNode: IRNode,
-    parentStack: IRNode[]
-): t.Expression {
-    if (t.isIdentifier(expression)) {
-        if (isComponentProp(expression, irNode, parentStack)) {
-            return t.memberExpression(t.identifier(TEMPLATE_PARAMS.INSTANCE), expression);
-        } else {
-            return expression;
-        }
+export default class Scope {
+    private readonly scopes: t.Identifier[][] = [];
+
+    beginScope() {
+        this.scopes.push([]);
     }
 
-    walk(expression, {
-        leave(node, parent) {
-            if (
-                parent !== null &&
-                t.isIdentifier(node) &&
-                t.isMemberExpression(parent) &&
-                parent.object === node &&
-                isComponentProp(node, irNode, parentStack)
-            ) {
-                this.replace(t.memberExpression(t.identifier(TEMPLATE_PARAMS.INSTANCE), node));
-            }
-        },
-    });
+    endScope() {
+        this.scopes.pop();
+    }
 
-    return expression;
+    peek() {
+        return this.scopes[this.scopes.length - 1];
+    }
+
+    declare(identifier: t.Identifier) {
+        this.peek().push(identifier);
+    }
+
+    resolve(identifier: t.Identifier) {
+        for (const id of this.scopes.flat()) {
+            if (identifier.name === id.name) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    /**
+     * Bind the passed expression to the component instance. It applies the following transformation to the expression:
+     * - {value} --> {$cmp.value}
+     * - {value[index]} --> {$cmp.value[$cmp.index]}
+     */
+    bindExpression(expression: Expression | Literal): t.Expression {
+        if (t.isIdentifier(expression)) {
+            if (this.resolve(expression)) {
+                return t.memberExpression(t.identifier(TEMPLATE_PARAMS.INSTANCE), expression);
+            } else {
+                return expression;
+            }
+        }
+
+        const scope = this;
+        walk(expression, {
+            leave(node, parent) {
+                if (
+                    parent !== null &&
+                    t.isIdentifier(node) &&
+                    t.isMemberExpression(parent) &&
+                    parent.object === node &&
+                    scope.resolve(node)
+                ) {
+                    this.replace(t.memberExpression(t.identifier(TEMPLATE_PARAMS.INSTANCE), node));
+                }
+            },
+        });
+
+        return expression as t.Expression;
+    }
 }
