--- packages/@lwc/template-compiler/src/parser-next/index.ts
@@ -1,1039 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: MIT
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
- */
-import * as parse5 from 'parse5';
-import { hasOwnProperty } from '@lwc/shared';
-
-import { ParserDiagnostics } from '@lwc/errors';
-import { cleanTextNode, decodeTextContent, parseHTML } from './html';
-
-import {
-    attributeName,
-    isAttribute,
-    isProhibitedIsAttribute,
-    isTabIndexAttribute,
-    isValidHTMLAttribute,
-    isValidTabIndexAttributeValue,
-    normalizeAttributeValue,
-    ParsedAttribute,
-    parseTagName,
-} from './attribute';
-
-import { isExpression, parseExpression, parseIdentifier } from './expression';
-
-import * as t from '../shared-next/estree';
-import * as parse5Utils from '../shared-next/parse5';
-import * as ir from '../shared-next/ir';
-import {
-    TemplateParseResult,
-    Attribute,
-    LWCDirectiveRenderMode,
-    Element,
-    Component,
-    ForEach,
-    Identifier,
-    Literal,
-    Expression,
-    ForOf,
-    Slot,
-    Text,
-    Comment,
-    Root,
-    LWCDirectiveDomMode,
-    SourceLocation,
-    ParentWrapper,
-    ParentNode,
-    ChildNode,
-    BaseElement,
-} from '../shared-next/types';
-
-import ParserCtx from './parser';
-
-import State from '../state';
-
-import {
-    DASHED_TAGNAME_ELEMENT_SET,
-    DISALLOWED_HTML_TAGS,
-    DISALLOWED_MATHML_TAGS,
-    EVENT_HANDLER_NAME_RE,
-    EVENT_HANDLER_RE,
-    EXPRESSION_RE,
-    HTML_NAMESPACE_URI,
-    IF_RE,
-    ITERATOR_RE,
-    KNOWN_HTML_ELEMENTS,
-    LWC_DIRECTIVE_SET,
-    LWC_DIRECTIVES,
-    LWC_RE,
-    MATHML_NAMESPACE_URI,
-    ROOT_TEMPLATE_DIRECTIVES,
-    ROOT_TEMPLATE_DIRECTIVES_SET,
-    SUPPORTED_SVG_TAGS,
-    SVG_NAMESPACE_URI,
-    VALID_IF_MODIFIER,
-    VOID_ELEMENT_SET,
-    FOR_DIRECTIVES,
-} from './constants';
-
-function attributeExpressionReferencesForOfIndex(attribute: Attribute, forOf: ForOf): boolean {
-    const { value } = attribute;
-    // if not an expression, it is not referencing iterator index
-    if (!t.isMemberExpression(value)) {
-        return false;
-    }
-
-    const { object, property } = value;
-    if (!t.isIdentifier(object) || !t.isIdentifier(property)) {
-        return false;
-    }
-
-    if (forOf.iterator.name !== object.name) {
-        return false;
-    }
-
-    return property.name === 'index';
-}
-
-function attributeExpressionReferencesForEachIndex(
-    attribute: Attribute,
-    forEach: ForEach
-): boolean {
-    const { index } = forEach;
-    const { value } = attribute;
-
-    // No index defined on foreach
-    if (!index || !t.isIdentifier(index) || !t.isIdentifier(value)) {
-        return false;
-    }
-
-    return index.name === value.name;
-}
-
-export default function parse(source: string, state: State): TemplateParseResult {
-    const ctx = new ParserCtx(source, state.config);
-    const fragment = parseHTML(ctx, source);
-
-    if (ctx.warnings.length) {
-        return { warnings: ctx.warnings };
-    }
-
-    const root = ctx.withErrorRecovery(() => {
-        const templateRoot = getTemplateRoot(ctx, fragment);
-        return parseRoot(ctx, templateRoot);
-    });
-
-    return { root, warnings: ctx.warnings };
-}
-
-function parseRoot(ctx: ParserCtx, parse5Elm: parse5.Element): Root {
-    const parsedAttr = parseAttributes(ctx, parse5Elm);
-    const root = ir.root(parse5Elm);
-
-    applyRootLwcDirectives(ctx, root, parsedAttr);
-    ctx.setRootDirective(root);
-    validateRoot(ctx, root, parse5Elm);
-    parseChildren(ctx, ir.parentWrapper(root), parse5Elm);
-
-    return root;
-}
-
-function parseElement(
-    ctx: ParserCtx,
-    parse5Elm: parse5.Element,
-    parent: ParentWrapper
-): ParentNode & ChildNode {
-    const parsedAttr = parseAttributes(ctx, parse5Elm);
-    const { head, current } = parseElementNodes(ctx, parse5Elm, parent, parsedAttr);
-    // parseElementDirectives will always return an Element | Component | Slot
-    const element = current.node as Component | Element | Slot;
-
-    applyHandlers(ctx, element, parsedAttr);
-    applyKey(ctx, element, current, parent, parsedAttr);
-    applyLwcDirectives(ctx, element, parsedAttr);
-    applyAttributes(ctx, element, current, parsedAttr);
-
-    validateElement(ctx, element, parse5Elm);
-    validateAttributes(ctx, element, parsedAttr);
-    validateProperties(ctx, element);
-
-    parseChildren(ctx, current, parse5Elm);
-    validateChildren(ctx, element);
-
-    return head;
-}
-
-function parseElementNodes(
-    ctx: ParserCtx,
-    parse5Elm: parse5.Element,
-    parent: ParentWrapper,
-    parsedAttr: ParsedAttribute
-): { head: ParentNode & ChildNode; current: ParentWrapper } {
-    let current = parent;
-    const elmNodes: ParentNode[] = [];
-    const parsers = [parseForEach, parseIterator, parseIf, parseElementType];
-
-    for (const parser of parsers) {
-        const node = parser(ctx, parse5Elm, parsedAttr, current);
-        if (node) {
-            const prev = elmNodes[elmNodes.length - 1];
-            prev?.children.push(node);
-            elmNodes.push(node);
-            current = ir.parentWrapper(node, current);
-        }
-    }
-
-    // elmNodes will always have at least one value because parseElementType
-    // will return Element | Component | Slot
-    return { head: elmNodes[0] as ParentNode & ChildNode, current };
-}
-
-function parseElementType(
-    ctx: ParserCtx,
-    parse5Elm: parse5.Element,
-    parsedAttr: ParsedAttribute,
-    parent: ParentWrapper
-) {
-    const { tagName: tag } = parse5Elm;
-    const location = ir.parseElementSourceLocation(parse5Elm);
-    // Check if the element tag is a valid custom element name and is not part of known standard
-    // element name containing a dash.
-    let element: Element | Component | Slot;
-    if (tag === 'slot') {
-        element = parseSlot(ctx, location, parsedAttr, parent);
-    } else if (!tag.includes('-') || DASHED_TAGNAME_ELEMENT_SET.has(tag)) {
-        element = ir.element(parse5Elm);
-    } else {
-        element = ir.component(parse5Elm);
-    }
-
-    return element;
-}
-
-function parseChildren(ctx: ParserCtx, parent: ParentWrapper, parse5Parent: parse5.Element): void {
-    const parsedChildren: ChildNode[] = [];
-    const children = (parse5Utils.getTemplateContent(parse5Parent) ?? parse5Parent).childNodes;
-
-    for (const child of children) {
-        ctx.withErrorRecovery(() => {
-            if (parse5Utils.isElementNode(child)) {
-                const elmNode = parseElement(ctx, child, parent);
-                parsedChildren.push(elmNode);
-            } else if (parse5Utils.isTextNode(child)) {
-                const textNodes = parseText(ctx, child);
-                parsedChildren.push(...textNodes);
-            } else if (parse5Utils.isCommentNode(child)) {
-                const commentNode = parseComment(child);
-                parsedChildren.push(commentNode);
-            }
-        });
-    }
-
-    parent.node.children = parsedChildren;
-}
-
-function parseText(ctx: ParserCtx, node: parse5.TextNode): Text[] {
-    const parsedTextNodes: Text[] = [];
-
-    // Extract the raw source to avoid HTML entity decoding done by parse5
-    const location = node.sourceCodeLocation!;
-    const rawText = cleanTextNode(ctx.getSource(location.startOffset, location.endOffset));
-
-    if (!rawText.trim().length) {
-        return parsedTextNodes;
-    }
-
-    // Split the text node content arround expression and create node for each
-    const tokenizedContent = rawText.split(EXPRESSION_RE);
-
-    for (const token of tokenizedContent) {
-        // Don't create nodes for emtpy strings
-        if (!token.length) {
-            continue;
-        }
-
-        let value: Expression | Literal;
-        if (isExpression(token)) {
-            value = parseExpression(ctx, token, ir.parseSourceLocation(location));
-        } else {
-            value = ir.literal(decodeTextContent(token));
-        }
-
-        parsedTextNodes.push(ir.text(node, value));
-    }
-
-    return parsedTextNodes;
-}
-
-function parseComment(node: parse5.CommentNode): Comment {
-    const value = decodeTextContent(node.data);
-    return ir.comment(node, value);
-}
-
-function getTemplateRoot(
-    ctx: ParserCtx,
-    documentFragment: parse5.DocumentFragment
-): parse5.Element {
-    // Filter all the empty text nodes
-    const validRoots = documentFragment.childNodes.filter(
-        (child) =>
-            parse5Utils.isElementNode(child) ||
-            (parse5Utils.isTextNode(child) && child.value.trim().length)
-    );
-
-    if (validRoots.length > 1) {
-        const duplicateRoot = validRoots[1].sourceCodeLocation!;
-        ctx.throwAtLocation(
-            ParserDiagnostics.MULTIPLE_ROOTS_FOUND,
-            ir.parseSourceLocation(duplicateRoot)
-        );
-    }
-
-    const [root] = validRoots;
-
-    if (!root || !parse5Utils.isElementNode(root)) {
-        ctx.throwAtLocation(ParserDiagnostics.MISSING_ROOT_TEMPLATE_TAG);
-    }
-
-    return root;
-}
-
-function applyHandlers(
-    ctx: ParserCtx,
-    element: Component | Element | Slot,
-    parsedAttr: ParsedAttribute
-) {
-    let eventHandlerAttribute;
-    while ((eventHandlerAttribute = parsedAttr.pick(EVENT_HANDLER_RE))) {
-        const { name } = eventHandlerAttribute;
-
-        if (!ir.isExpression(eventHandlerAttribute.value)) {
-            ctx.throwOnNode(
-                ParserDiagnostics.EVENT_HANDLER_SHOULD_BE_EXPRESSION,
-                eventHandlerAttribute
-            );
-        }
-
-        if (!name.match(EVENT_HANDLER_NAME_RE)) {
-            ctx.throwOnNode(ParserDiagnostics.INVALID_EVENT_NAME, eventHandlerAttribute, [name]);
-        }
-
-        // Strip the `on` prefix from the event handler name
-        const eventName = name.slice(2);
-        const listener = ir.eventListener(
-            eventName,
-            eventHandlerAttribute.value,
-            eventHandlerAttribute.location
-        );
-        element.listeners.push(listener);
-    }
-}
-
-function parseIf(ctx: ParserCtx, parse5Elm: parse5.Element, parsedAttr: ParsedAttribute) {
-    const ifAttribute = parsedAttr.pick(IF_RE);
-    if (!ifAttribute) {
-        return;
-    }
-
-    if (!ir.isExpression(ifAttribute.value)) {
-        ctx.throwOnNode(ParserDiagnostics.IF_DIRECTIVE_SHOULD_BE_EXPRESSION, ifAttribute);
-    }
-
-    const [, modifier] = ifAttribute.name.split(':');
-    if (!VALID_IF_MODIFIER.has(modifier)) {
-        ctx.throwOnNode(ParserDiagnostics.UNEXPECTED_IF_MODIFIER, ifAttribute, [modifier]);
-    }
-
-    return ir.ifBlock(parse5Elm.tagName, ifAttribute.location, modifier, ifAttribute.value);
-}
-
-function applyRootLwcDirectives(ctx: ParserCtx, root: Root, parsedAttr: ParsedAttribute) {
-    const lwcAttribute = parsedAttr.get(LWC_RE);
-    if (!lwcAttribute) {
-        return;
-    }
-
-    if (
-        !LWC_DIRECTIVE_SET.has(lwcAttribute.name) &&
-        !ROOT_TEMPLATE_DIRECTIVES_SET.has(lwcAttribute.name)
-    ) {
-        ctx.throwOnNode(ParserDiagnostics.UNKNOWN_LWC_DIRECTIVE, root, [
-            lwcAttribute.name,
-            `<${root.name}>`,
-        ]);
-    }
-
-    applyLwcRenderModeDirective(ctx, root, parsedAttr);
-    applyLwcPreserveCommentsDirective(ctx, root, parsedAttr);
-}
-
-function applyLwcRenderModeDirective(ctx: ParserCtx, root: Root, parsedAttr: ParsedAttribute) {
-    const lwcRenderModeAttribute = parsedAttr.get(ROOT_TEMPLATE_DIRECTIVES.RENDER_MODE);
-    if (!lwcRenderModeAttribute) {
-        return;
-    }
-
-    const { value: renderDomAttr } = lwcRenderModeAttribute;
-
-    if (
-        !ir.isStringLiteral(renderDomAttr) ||
-        (renderDomAttr.value !== LWCDirectiveRenderMode.Shadow &&
-            renderDomAttr.value !== LWCDirectiveRenderMode.Light)
-    ) {
-        ctx.throwOnNode(ParserDiagnostics.LWC_RENDER_MODE_INVALID_VALUE, root);
-    }
-
-    const value = ir.renderModeDirective(renderDomAttr.value, lwcRenderModeAttribute.location);
-    const directives = root.directives || (root.directives = []);
-    directives.push(value);
-}
-
-function applyLwcPreserveCommentsDirective(
-    ctx: ParserCtx,
-    root: Root,
-    parsedAttr: ParsedAttribute
-) {
-    const lwcPreserveCommentAttribute = parsedAttr.get(ROOT_TEMPLATE_DIRECTIVES.PRESERVE_COMMENTS);
-    if (!lwcPreserveCommentAttribute) {
-        return;
-    }
-
-    const { value: lwcPreserveCommentsAttr } = lwcPreserveCommentAttribute;
-
-    if (!ir.isBooleanLiteral(lwcPreserveCommentsAttr)) {
-        ctx.throwOnNode(ParserDiagnostics.UNKNOWN_LWC_DIRECTIVE, root, [
-            ROOT_TEMPLATE_DIRECTIVES.RENDER_MODE,
-            `<${root.name}>`,
-        ]);
-    }
-
-    const value = ir.preserveCommentsDirective(
-        lwcPreserveCommentsAttr.value,
-        lwcPreserveCommentAttribute.location
-    );
-    const directives = root.directives || (root.directives = []);
-    directives.push(value);
-}
-
-function applyLwcDirectives(
-    ctx: ParserCtx,
-    element: Element | Component | Slot,
-    parsedAttr: ParsedAttribute
-) {
-    const lwcAttribute = parsedAttr.get(LWC_RE);
-    if (!lwcAttribute) {
-        return;
-    }
-
-    if (
-        !LWC_DIRECTIVE_SET.has(lwcAttribute.name) &&
-        !ROOT_TEMPLATE_DIRECTIVES_SET.has(lwcAttribute.name)
-    ) {
-        ctx.throwOnNode(ParserDiagnostics.UNKNOWN_LWC_DIRECTIVE, element, [
-            lwcAttribute.name,
-            `<${element.name}>`,
-        ]);
-    }
-
-    // Should not allow render mode or preserve comments on non root nodes
-    if (parsedAttr.get(ROOT_TEMPLATE_DIRECTIVES.RENDER_MODE)) {
-        ctx.throwOnNode(ParserDiagnostics.UNKNOWN_LWC_DIRECTIVE, element, [
-            ROOT_TEMPLATE_DIRECTIVES.RENDER_MODE,
-            `<${element.name}>`,
-        ]);
-    }
-
-    if (parsedAttr.get(ROOT_TEMPLATE_DIRECTIVES.PRESERVE_COMMENTS)) {
-        ctx.throwOnNode(ParserDiagnostics.UNKNOWN_LWC_DIRECTIVE, element, [
-            ROOT_TEMPLATE_DIRECTIVES.RENDER_MODE,
-            `<${element.name}>`,
-        ]);
-    }
-
-    applyLwcDynamicDirective(ctx, element, parsedAttr);
-    applyLwcDomDirective(ctx, element, parsedAttr);
-}
-
-function applyLwcDynamicDirective(
-    ctx: ParserCtx,
-    element: Element | Component | Slot,
-    parsedAttr: ParsedAttribute
-) {
-    const tag = parseTagName(element);
-
-    const lwcDynamicAttribute = parsedAttr.pick(LWC_DIRECTIVES.DYNAMIC);
-    if (!lwcDynamicAttribute) {
-        return;
-    }
-
-    if (!ctx.config.experimentalDynamicDirective) {
-        ctx.throwOnNode(ParserDiagnostics.INVALID_OPTS_LWC_DYNAMIC, element, [`<${tag}>`]);
-    }
-
-    if (!ir.isComponent(element)) {
-        ctx.throwOnNode(ParserDiagnostics.INVALID_LWC_DYNAMIC_ON_NATIVE_ELEMENT, element, [
-            `<${tag}>`,
-        ]);
-    }
-
-    const { value: lwcDynamicAttr } = lwcDynamicAttribute;
-    if (!ir.isExpression(lwcDynamicAttr)) {
-        ctx.throwOnNode(ParserDiagnostics.INVALID_LWC_DYNAMIC_LITERAL_PROP, element, [`<${tag}>`]);
-    }
-
-    const directives = element.directives || (element.directives = []);
-    directives.push(ir.dynamicDirective(lwcDynamicAttr, lwcDynamicAttr.location));
-}
-
-function applyLwcDomDirective(
-    ctx: ParserCtx,
-    element: Element | Component | Slot,
-    parsedAttr: ParsedAttribute
-) {
-    const tag = parseTagName(element);
-
-    const lwcDomAttribute = parsedAttr.pick(LWC_DIRECTIVES.DOM);
-    if (!lwcDomAttribute) {
-        return;
-    }
-
-    if (ctx.renderMode === LWCDirectiveRenderMode.Light) {
-        ctx.throwOnNode(ParserDiagnostics.LWC_DOM_INVALID_IN_LIGHT_DOM, element, [`<${tag}>`]);
-    }
-
-    if (ir.isComponent(element)) {
-        ctx.throwOnNode(ParserDiagnostics.LWC_DOM_INVALID_CUSTOM_ELEMENT, element, [`<${tag}>`]);
-    }
-
-    if (ir.isSlot(element)) {
-        ctx.throwOnNode(ParserDiagnostics.LWC_DOM_INVALID_SLOT_ELEMENT, element);
-    }
-
-    const { value: lwcDomAttr } = lwcDomAttribute;
-
-    if (
-        !ir.isStringLiteral(lwcDomAttr) ||
-        hasOwnProperty.call(LWCDirectiveDomMode, lwcDomAttr.value) === false
-    ) {
-        const possibleValues = Object.keys(LWCDirectiveDomMode)
-            .map((value) => `"${value}"`)
-            .join(', or ');
-        ctx.throwOnNode(ParserDiagnostics.LWC_DOM_INVALID_VALUE, element, [possibleValues]);
-    }
-
-    const directives = element.directives || (element.directives = []);
-    directives.push(ir.domDirective(lwcDomAttr.value, lwcDomAttribute.location));
-}
-
-function parseForEach(ctx: ParserCtx, parse5Elm: parse5.Element, parsedAttr: ParsedAttribute) {
-    const forEachAttribute = parsedAttr.pick(FOR_DIRECTIVES.FOR_EACH);
-    const forItemAttribute = parsedAttr.pick(FOR_DIRECTIVES.FOR_ITEM);
-    const forIndex = parsedAttr.pick(FOR_DIRECTIVES.FOR_INDEX);
-
-    if (forEachAttribute && forItemAttribute) {
-        if (!ir.isExpression(forEachAttribute.value)) {
-            ctx.throwOnNode(
-                ParserDiagnostics.FOR_EACH_DIRECTIVE_SHOULD_BE_EXPRESSION,
-                forEachAttribute
-            );
-        }
-
-        const forItemValue = forItemAttribute.value;
-        if (!ir.isStringLiteral(forItemValue)) {
-            ctx.throwOnNode(
-                ParserDiagnostics.FOR_ITEM_DIRECTIVE_SHOULD_BE_STRING,
-                forItemAttribute
-            );
-        }
-
-        const item = parseIdentifier(ctx, forItemValue.value, forItemAttribute.location);
-
-        let index: Identifier | undefined;
-        if (forIndex) {
-            const forIndexValue = forIndex.value;
-            if (!ir.isStringLiteral(forIndexValue)) {
-                ctx.throwOnNode(ParserDiagnostics.FOR_INDEX_DIRECTIVE_SHOULD_BE_STRING, forIndex);
-            }
-
-            index = parseIdentifier(ctx, forIndexValue.value, forIndex.location);
-        }
-        return ir.forEach(
-            parse5Elm.tagName,
-            forEachAttribute.value,
-            forEachAttribute.location,
-            item,
-            index
-        );
-    } else if (forEachAttribute || forItemAttribute) {
-        const location = ir.parseElementSourceLocation(parse5Elm);
-        ctx.throwAtLocation(
-            ParserDiagnostics.FOR_EACH_AND_FOR_ITEM_DIRECTIVES_SHOULD_BE_TOGETHER,
-            location
-        );
-    }
-}
-
-function parseIterator(
-    ctx: ParserCtx,
-    parse5Elm: parse5.Element,
-    parsedAttr: ParsedAttribute,
-    current: ParentWrapper
-) {
-    const iteratorExpression = parsedAttr.pick(ITERATOR_RE);
-    if (!iteratorExpression) {
-        return;
-    }
-
-    if (ir.isForEach(current.node)) {
-        const location = ir.parseElementSourceLocation(parse5Elm);
-        ctx.throwAtLocation(ParserDiagnostics.INVALID_FOR_EACH_WITH_ITERATOR, location, [
-            iteratorExpression.name,
-        ]);
-    }
-
-    const iteratorAttributeName = iteratorExpression.name;
-    const [, iteratorName] = iteratorAttributeName.split(':');
-
-    if (!ir.isExpression(iteratorExpression.value)) {
-        ctx.throwOnNode(ParserDiagnostics.DIRECTIVE_SHOULD_BE_EXPRESSION, iteratorExpression, [
-            iteratorExpression.name,
-        ]);
-    }
-
-    const iterator = parseIdentifier(ctx, iteratorName, iteratorExpression.location);
-
-    return ir.forOf(
-        parse5Elm.tagName,
-        iteratorExpression.value,
-        iterator,
-        iteratorExpression.location
-    );
-}
-
-function applyKey(
-    ctx: ParserCtx,
-    element: Component | Element | Slot,
-    current: ParentWrapper,
-    parent: ParentWrapper,
-    parsedAttr: ParsedAttribute
-) {
-    const tag = parseTagName(element);
-    const keyAttribute = parsedAttr.pick('key');
-
-    if (keyAttribute) {
-        if (!ir.isExpression(keyAttribute.value)) {
-            ctx.throwOnNode(ParserDiagnostics.KEY_ATTRIBUTE_SHOULD_BE_EXPRESSION, keyAttribute);
-        }
-
-        const forOfParent = getIteratorParent(ctx, parent);
-        const forEachParent = getForEachParent(ctx, current);
-
-        if (forOfParent) {
-            if (attributeExpressionReferencesForOfIndex(keyAttribute, forOfParent)) {
-                ctx.throwOnNode(
-                    ParserDiagnostics.KEY_SHOULDNT_REFERENCE_ITERATOR_INDEX,
-                    keyAttribute,
-                    [tag]
-                );
-            }
-        } else if (forEachParent) {
-            if (attributeExpressionReferencesForEachIndex(keyAttribute, forEachParent)) {
-                const name = 'name' in keyAttribute.value && keyAttribute.value.name;
-                ctx.throwOnNode(
-                    ParserDiagnostics.KEY_SHOULDNT_REFERENCE_FOR_EACH_INDEX,
-                    keyAttribute,
-                    [tag, name]
-                );
-            }
-        }
-
-        const forKey = ir.keyDirective(keyAttribute.value, keyAttribute.location);
-        const directive = element.directives || (element.directives = []);
-        directive.push(forKey);
-    } else if (isInIteratorElement(ctx, current, parent) && !ir.isTemplate(element)) {
-        ctx.throwOnNode(ParserDiagnostics.MISSING_KEY_IN_ITERATOR, element, [tag]);
-    }
-}
-
-function parseSlot(
-    ctx: ParserCtx,
-    elementLocation: SourceLocation,
-    parsedAttr: ParsedAttribute,
-    parent: ParentWrapper
-): Slot {
-    if (ir.isForBlock(parent.node) || ir.isIfBlock(parent.node)) {
-        ctx.throwAtLocation(ParserDiagnostics.SLOT_TAG_CANNOT_HAVE_DIRECTIVES, elementLocation);
-    }
-
-    // Can't handle slots in applySlot because it would be too late for class and style attrs
-    if (ctx.renderMode === LWCDirectiveRenderMode.Light) {
-        const invalidAttrs = parsedAttr
-            .getAttributes()
-            .filter(({ name }) => name !== 'name')
-            .map(({ name }) => name);
-
-        if (invalidAttrs.length > 0) {
-            ctx.throwAtLocation(
-                ParserDiagnostics.LWC_LIGHT_SLOT_INVALID_ATTRIBUTES,
-                elementLocation,
-                [invalidAttrs.join(',')]
-            );
-        }
-    }
-
-    // Default slot have empty string name
-    let name = '';
-
-    const nameAttribute = parsedAttr.get('name');
-    if (nameAttribute) {
-        if (ir.isExpression(nameAttribute.value)) {
-            ctx.throwOnNode(ParserDiagnostics.NAME_ON_SLOT_CANNOT_BE_EXPRESSION, nameAttribute);
-        } else if (ir.isStringLiteral(nameAttribute.value)) {
-            name = nameAttribute.value.value;
-        }
-    }
-
-    const alreadySeen = ctx.seenSlots.has(name);
-    ctx.seenSlots.add(name);
-
-    if (alreadySeen) {
-        ctx.warnAtLocation(
-            ParserDiagnostics.NO_DUPLICATE_SLOTS,
-            [name === '' ? 'default' : `name="${name}"`],
-            elementLocation
-        );
-    } else if (isInIteration(ctx, parent)) {
-        ctx.warnAtLocation(
-            ParserDiagnostics.NO_SLOTS_IN_ITERATOR,
-            [name === '' ? 'default' : `name="${name}"`],
-            elementLocation
-        );
-    }
-
-    return ir.slot(name, elementLocation);
-}
-
-function applyAttributes(
-    ctx: ParserCtx,
-    element: Element | Component | Slot,
-    current: ParentWrapper,
-    parsedAttr: ParsedAttribute
-) {
-    const tag = parseTagName(element);
-    const attributes = parsedAttr.getAttributes();
-
-    for (const attr of attributes) {
-        const { name } = attr;
-
-        if (!isValidHTMLAttribute(tag, name)) {
-            ctx.warnOnNode(ParserDiagnostics.INVALID_HTML_ATTRIBUTE, attr, [name, tag]);
-        }
-
-        if (name.match(/[^a-z0-9]$/)) {
-            ctx.throwOnNode(
-                ParserDiagnostics.ATTRIBUTE_NAME_MUST_END_WITH_ALPHA_NUMERIC_CHARACTER,
-                attr,
-                [name, tag]
-            );
-        }
-
-        if (!/^-*[a-z]/.test(name)) {
-            ctx.throwOnNode(
-                ParserDiagnostics.ATTRIBUTE_NAME_MUST_START_WITH_ALPHABETIC_OR_HYPHEN_CHARACTER,
-                attr,
-                [name, tag]
-            );
-        }
-
-        // disallow attr name which combines underscore character with special character.
-        // We normalize camel-cased names with underscores caMel -> ca-mel; thus sanitization.
-        if (name.match(/_[^a-z0-9]|[^a-z0-9]_/)) {
-            ctx.throwOnNode(
-                ParserDiagnostics.ATTRIBUTE_NAME_CANNOT_COMBINE_UNDERSCORE_WITH_SPECIAL_CHARS,
-                attr,
-                [name, tag]
-            );
-        }
-
-        if (ir.isStringLiteral(attr.value)) {
-            if (name === 'id') {
-                const { value } = attr.value;
-
-                if (/\s+/.test(value)) {
-                    ctx.throwOnNode(ParserDiagnostics.INVALID_ID_ATTRIBUTE, attr, [value]);
-                }
-
-                if (isInIteration(ctx, current)) {
-                    ctx.throwOnNode(ParserDiagnostics.INVALID_STATIC_ID_IN_ITERATION, attr, [
-                        value,
-                    ]);
-                }
-
-                if (ctx.seenIds.has(value)) {
-                    ctx.throwOnNode(ParserDiagnostics.DUPLICATE_ID_FOUND, attr, [value]);
-                } else {
-                    ctx.seenIds.add(value);
-                }
-            }
-        }
-
-        // Prevent usage of the slot attribute with expression.
-        if (name === 'slot' && ir.isExpression(attr.value)) {
-            ctx.throwOnNode(ParserDiagnostics.SLOT_ATTRIBUTE_CANNOT_BE_EXPRESSION, attr);
-        }
-
-        // the if branch handles
-        // 1. All attributes for standard elements except 1 case are handled as attributes
-        // 2. For custom elements, only key, slot and data are handled as attributes, rest as properties
-        if (isAttribute(element, name)) {
-            element.attributes.push(attr);
-        } else {
-            element.properties.push(ir.property(name, attr.value, attr.location));
-
-            parsedAttr.pick(name);
-        }
-    }
-}
-
-function validateRoot(ctx: ParserCtx, root: Root, parse5Elm: parse5.Element) {
-    const { tagName: tag } = parse5Elm;
-
-    if (!ir.isTemplate(root)) {
-        ctx.throwOnNode(ParserDiagnostics.ROOT_TAG_SHOULD_BE_TEMPLATE, root, [tag]);
-    }
-
-    const rootHasUnknownAttributes = parse5Elm.attrs.some(
-        ({ name }) => !ROOT_TEMPLATE_DIRECTIVES_SET.has(name)
-    );
-
-    if (rootHasUnknownAttributes) {
-        ctx.throwOnNode(ParserDiagnostics.ROOT_TEMPLATE_HAS_UNKNOWN_ATTRIBUTES, root);
-    }
-
-    // Check if a non-void element has a matching closing tag.
-    //
-    // Note: Parse5 currently fails to collect end tag location for element with a tag name
-    // containing an upper case character (inikulin/parse5#352).
-    const location = ir.parseElementLocation(parse5Elm);
-    const hasClosingTag = Boolean(location.endTag);
-    const isVoidElement = VOID_ELEMENT_SET.has(tag);
-    if (!isVoidElement && !hasClosingTag && tag === tag.toLocaleLowerCase()) {
-        ctx.throwOnNode(ParserDiagnostics.NO_MATCHING_CLOSING_TAGS, root, [tag]);
-    }
-}
-
-function validateElement(
-    ctx: ParserCtx,
-    element: Element | Component | Slot,
-    parse5Elm: parse5.Element
-) {
-    const { tagName: tag, namespaceURI: namespace } = parse5Elm;
-    const elementLocation = ir.parseElementLocation(parse5Elm);
-
-    // Check if a non-void element has a matching closing tag.
-    //
-    // Note: Parse5 currently fails to collect end tag location for element with a tag name
-    // containing an upper case character (inikulin/parse5#352).
-    const hasClosingTag = Boolean(elementLocation.endTag);
-    const isVoidElement = VOID_ELEMENT_SET.has(tag);
-    if (!isVoidElement && !hasClosingTag && tag === tag.toLocaleLowerCase()) {
-        ctx.throwOnNode(ParserDiagnostics.NO_MATCHING_CLOSING_TAGS, element, [tag]);
-    }
-
-    if (tag === 'style' && namespace === HTML_NAMESPACE_URI) {
-        ctx.throwOnNode(ParserDiagnostics.STYLE_TAG_NOT_ALLOWED_IN_TEMPLATE, element);
-    } else if (ir.isTemplate(element)) {
-        // We check if the template element has some modifier applied to it. Directly checking if one of the
-        // IRElement property is impossible. For example when an error occurs during the parsing of the if
-        // expression, the `element.if` property remains undefined. It would results in 2 warnings instead of 1:
-        //      - Invalid if expression
-        //      - Unexpected template element
-        //
-        // Checking if the original HTMLElement has some attributes applied is a good enough for now.
-        const hasAttributes = parse5Elm.attrs.length !== 0;
-        // if (!isRoot && !hasAttributes) {
-        if (!hasAttributes) {
-            ctx.throwOnNode(ParserDiagnostics.NO_DIRECTIVE_FOUND_ON_TEMPLATE, element);
-        }
-    } else {
-        const isNotAllowedHtmlTag = DISALLOWED_HTML_TAGS.has(tag);
-        if (namespace === HTML_NAMESPACE_URI && isNotAllowedHtmlTag) {
-            ctx.throwOnNode(ParserDiagnostics.FORBIDDEN_TAG_ON_TEMPLATE, element, [tag]);
-        }
-
-        const isNotAllowedSvgTag = !SUPPORTED_SVG_TAGS.has(tag);
-        if (namespace === SVG_NAMESPACE_URI && isNotAllowedSvgTag) {
-            ctx.throwOnNode(ParserDiagnostics.FORBIDDEN_SVG_NAMESPACE_IN_TEMPLATE, element, [tag]);
-        }
-
-        const isNotAllowedMathMlTag = DISALLOWED_MATHML_TAGS.has(tag);
-        if (namespace === MATHML_NAMESPACE_URI && isNotAllowedMathMlTag) {
-            ctx.throwOnNode(ParserDiagnostics.FORBIDDEN_MATHML_NAMESPACE_IN_TEMPLATE, element, [
-                tag,
-            ]);
-        }
-
-        const isKnownTag =
-            ir.isComponent(element) ||
-            KNOWN_HTML_ELEMENTS.has(tag) ||
-            SUPPORTED_SVG_TAGS.has(tag) ||
-            DASHED_TAGNAME_ELEMENT_SET.has(tag);
-
-        if (!isKnownTag) {
-            ctx.warnOnNode(ParserDiagnostics.UNKNOWN_HTML_TAG_IN_TEMPLATE, element, [tag]);
-        }
-    }
-}
-
-function validateChildren(ctx: ParserCtx, element: Element | Component | Slot) {
-    const effectiveChildren = ctx.preserveComments
-        ? element.children
-        : element.children.filter((child) => !ir.isCommentNode(child));
-    const hasDomDirective = ir.getDomDirective(element);
-    if (hasDomDirective && effectiveChildren.length > 0) {
-        ctx.throwOnNode(ParserDiagnostics.LWC_DOM_INVALID_CONTENTS, element);
-    }
-}
-
-function validateAttributes(
-    ctx: ParserCtx,
-    element: Element | Component | Slot,
-    parsedAttr: ParsedAttribute
-) {
-    const tag = parseTagName(element);
-    const attributes = parsedAttr.getAttributes();
-
-    for (const attr of attributes) {
-        const { name: attrName, value: attrVal } = attr;
-
-        if (isProhibitedIsAttribute(attrName)) {
-            ctx.throwOnNode(ParserDiagnostics.IS_ATTRIBUTE_NOT_SUPPORTED, element, [attrName, tag]);
-        }
-
-        if (isTabIndexAttribute(attrName)) {
-            if (!ir.isExpression(attrVal) && !isValidTabIndexAttributeValue(attrVal.value)) {
-                ctx.throwOnNode(ParserDiagnostics.INVALID_TABINDEX_ATTRIBUTE, element);
-            }
-        }
-
-        // TODO [#1136]: once the template compiler emits the element namespace information to the engine we should
-        // restrict the validation of the "srcdoc" attribute on the "iframe" element only if this element is
-        // part of the HTML namespace.
-        if (tag === 'iframe' && attrName === 'srcdoc') {
-            ctx.throwOnNode(ParserDiagnostics.FORBIDDEN_IFRAME_SRCDOC_ATTRIBUTE, element);
-        }
-    }
-}
-
-function validateProperties(ctx: ParserCtx, element: BaseElement) {
-    const tag = parseTagName(element);
-
-    for (const prop of element.properties) {
-        const { name, value } = prop;
-
-        if (isProhibitedIsAttribute(name)) {
-            ctx.throwOnNode(ParserDiagnostics.IS_ATTRIBUTE_NOT_SUPPORTED, element, [name, tag]);
-        }
-
-        if (
-            // tabindex is transformed to tabIndex for properties
-            isTabIndexAttribute(name.toLowerCase()) &&
-            !ir.isExpression(value) &&
-            !isValidTabIndexAttributeValue(value.value)
-        ) {
-            ctx.throwOnNode(ParserDiagnostics.INVALID_TABINDEX_ATTRIBUTE, element);
-        }
-    }
-}
-
-function parseAttributes(ctx: ParserCtx, parse5Elm: parse5.Element): ParsedAttribute {
-    const parsedAttrs = new ParsedAttribute();
-    const { attrs: attributes, tagName } = parse5Elm;
-    const { attrs: attrLocations } = ir.parseElementLocation(parse5Elm);
-
-    for (const attr of attributes) {
-        parsedAttrs.append(getTemplateAttribute(ctx, tagName, attrLocations, attr));
-    }
-
-    return parsedAttrs;
-}
-
-function getTemplateAttribute(
-    ctx: ParserCtx,
-    tag: string,
-    attrLocations: parse5.AttributesLocation,
-    attribute: parse5.Attribute
-): Attribute {
-    const name = attributeName(attribute);
-
-    // Convert attribute name to lowercase because the location map keys follow the algorithm defined in the spec
-    // https://wicg.github.io/controls-list/html-output/multipage/syntax.html#attribute-name-state
-    const rawLocation = attrLocations[name.toLowerCase()];
-    const rawAttribute = ctx.getSource(rawLocation.startOffset, rawLocation.endOffset);
-
-    const location = ir.parseSourceLocation(rawLocation);
-    // parse5 automatically converts the casing from camelcase to all lowercase. If the attribute name
-    // is not the same before and after the parsing, then the attribute name contains capital letters
-    if (!rawAttribute.startsWith(name)) {
-        ctx.throwAtLocation(ParserDiagnostics.INVALID_ATTRIBUTE_CASE, location, [
-            rawAttribute,
-            tag,
-        ]);
-    }
-
-    const isBooleanAttribute = !rawAttribute.includes('=');
-    const { value, escapedExpression } = normalizeAttributeValue(
-        ctx,
-        rawAttribute,
-        tag,
-        attribute,
-        location
-    );
-
-    let attrValue: Literal | Expression;
-    if (isExpression(value) && !escapedExpression) {
-        attrValue = parseExpression(ctx, value, location);
-    } else if (isBooleanAttribute) {
-        attrValue = ir.literal(true);
-    } else {
-        attrValue = ir.literal(value);
-    }
-
-    return ir.attribute(name, attrValue, location);
-}
-
-function isInIteration(ctx: ParserCtx, element: ParentWrapper): boolean {
-    return !!ctx.findAncestor({
-        current: element,
-        predicate: (wrapper) => ir.isTemplate(wrapper.node) && ir.isForBlock(wrapper.node),
-    });
-}
-
-function getIteratorParent(ctx: ParserCtx, srcNode: ParentWrapper): ForOf | null {
-    const result = ctx.findAncestor({
-        current: srcNode,
-        predicate: (wrapper) => ir.isForOf(wrapper.node),
-        traversalCond: ({ current, parent }) =>
-            !ir.isBaseElement(parent.node) || ir.isTemplate(current.node),
-    });
-    return result ? (result as ForOf) : null;
-}
-
-function getForEachParent(ctx: ParserCtx, current: ParentWrapper): ForEach | null {
-    const result = ctx.findAncestor({
-        current,
-        predicate: (wrapper) => ir.isForEach(wrapper.node),
-        traversalCond: ({ parent }) => !ir.isBaseElement(parent.node) || ir.isTemplate(parent.node),
-    });
-    return result ? (result as ForEach) : null;
-}
-
-function isInIteratorElement(
-    ctx: ParserCtx,
-    current: ParentWrapper,
-    parent: ParentWrapper
-): boolean {
-    return !!(getIteratorParent(ctx, parent) || getForEachParent(ctx, current));
-}
